Lab 6 | data validations complete

+++ Class 7/lab/data/bands.js
@@ -1,6 +1,7 @@
 // This data file should export all functions using the ES6 standard as shown in the lecture code
 
 import { bands } from "../config/mongoCollections.js";
+import * as helper from "../helpers.js";
 import { ObjectId } from "mongodb";
 
 const create = async (
@@ -11,16 +12,67 @@ const create = async (
     groupMembers,
     yearBandWasFormed
 ) => {
+    if (
+        !helper.exists(
+            name,
+            genre,
+            website,
+            recordCompany,
+            groupMembers,
+            yearBandWasFormed
+        )
+    ) {
+        throw new Error("All fields are required");
+    }
+
+    if (!helper.isValidArray(genre, groupMembers)) {
+        throw new Error(
+            "genre and groupMembers must be a valid non-empty array."
+        );
+    }
+
+    if (
+        !helper.isValidString(
+            name,
+            ...genre,
+            ...groupMembers,
+            website,
+            recordCompany
+        )
+    ) {
+        throw new Error(
+            "Name, website, recordCompany, all genres and all group members must be non-empty strings"
+        );
+    }
+
+    if (!/^http:\/\/www\.[\w-]{5,}\.com$/.test(website)) {
+        throw new Error(
+            "Website must be a valid format and have at least 5 characters in-between http://www. and .com"
+        );
+    }
+
+    if (
+        typeof yearBandWasFormed !== "number" ||
+        isNaN(yearBandWasFormed) ||
+        yearBandWasFormed < 1900 ||
+        yearBandWasFormed > new Date().getFullYear()
+    ) {
+        throw new Error(
+            "Year band was formed must be a number between 1900 and the current year (2023)"
+        );
+    }
+
     let newBand = {
-        name: name,
-        genre: genre,
-        website: website,
-        recordCompany: recordCompany,
-        groupMembers: groupMembers,
+        name: name.trim(),
+        genre: genre.map((g) => g.trim()),
+        website: website.trim(),
+        recordCompany: recordCompany.trim(),
+        groupMembers: groupMembers.map((m) => m.trim()),
         yearBandWasFormed: yearBandWasFormed,
         albums: [],
         overallRating: 0,
     };
+
     const bandCollection = await bands();
     const insertInfo = await bandCollection.insertOne(newBand);
 
@@ -45,6 +97,11 @@ const getAll = async () => {
 };
 
 const get = async (id) => {
+    if (!helper.exists(id)) throw new Error("No id provided");
+    if (!helper.isValidString(id))
+        throw new Error("id must be a non-empty string");
+    if (!ObjectId.isValid(id)) throw new Error("Invalid ObjectId provided");
+
     const bandCollection = await bands();
     const band = await bandCollection.findOne({ _id: new ObjectId(id) });
     if (!band) throw new Error("No band with that id");
@@ -54,6 +111,11 @@ const get = async (id) => {
 };
 
 const remove = async (id) => {
+    if (!helper.exists(id)) throw new Error("No id provided");
+    if (!helper.isValidString(id))
+        throw new Error("id must be a non-empty string");
+    if (!ObjectId.isValid(id)) throw new Error("Invalid ObjectId provided");
+
     const bandCollection = await bands();
     const deletionInfo = await bandCollection.findOneAndDelete({
         _id: new ObjectId(id),
@@ -76,6 +138,61 @@ const update = async (
     groupMembers,
     yearBandWasFormed
 ) => {
+    if (!helper.exists(id)) throw new Error("No id provided");
+    if (!helper.isValidString(id))
+        throw new Error("id must be a non-empty string");
+    if (!ObjectId.isValid(id)) throw new Error("Invalid ObjectId provided");
+
+    if (
+        !helper.exists(
+            name,
+            genre,
+            website,
+            recordCompany,
+            groupMembers,
+            yearBandWasFormed
+        )
+    ) {
+        throw new Error("All fields are required");
+    }
+
+    if (!helper.isValidArray(genre, groupMembers)) {
+        throw new Error(
+            "genre and groupMembers must be a valid non-empty array."
+        );
+    }
+
+    if (
+        !helper.isValidString(
+            name,
+            ...genre,
+            ...groupMembers,
+            website,
+            recordCompany
+        )
+    ) {
+        throw new Error(
+            "Name, website, recordCompany, all genres and all group members must be non-empty strings"
+        );
+    }
+
+    if (!/^http:\/\/www\.[\w-]{5,}\.com$/.test(website)) {
+        throw new Error(
+            "Website must be a valid format and have at least 5 characters in-between http://www. and .com"
+        );
+    }
+
+    if (
+        typeof yearBandWasFormed !== "number" ||
+        isNaN(yearBandWasFormed) ||
+        yearBandWasFormed < 1900 ||
+        yearBandWasFormed > new Date().getFullYear()
+    ) {
+        throw new Error(
+            "Year band was formed must be a number between 1900 and the current year (2023)"
+        );
+    }
+
     const bandCollection = await bands();
     const filter = { _id: new ObjectId(id) };
     const updateDoc = {
@@ -89,7 +206,14 @@ const update = async (
         },
     };
 
+    const band = await get(id);
+    if (band.name.toLowerCase().trim() === newName.toLowerCase().trim()) {
+        throw new Error("New name cannot be the same as current name");
+    }
+
     const updateInfo = await bandCollection.updateOne(filter, updateDoc);
+    if (updateInfo.modifiedCount === 0) throw new Error("No band with that id");
+
     const updatedBand = await get(id);
     return updatedBand;
 };
